pyproject.toml
[tool.poetry]
name = "Auth Service"
version = "0.1.0"
authors = ["Barahlush <baraltiva@gmail.com>", "Polinavas95 <tatsuchan@mail.ru>"]
description = "Authentification service for the online movie database."
packages = [
    { include = "auth_service"}
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "0.21.1"
flask = "^2.2.2"
gevent = ">=22.10.2,<22.11.0"
wheel = "0.38.1"
gunicorn = ">=20.1.0,<20.2.0"
pytest = "^7.2.1"
sqlalchemy = ">=2.0.2,<2.1.0"
psycopg2-binary = ">=2.9.5,<2.10.0"
flask-sqlalchemy = ">=3.0.3,<3.1.0"
flask-migrate = ">=4.0.4,<4.1.0"
pydantic = ">=1.10.2,<1.11.0"

[tool.poetry.group.dev.dependencies]
blue = "*"
mypy = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
ruff = "*"
pre-commit = "*"

[tool.blue]
line-length = 79

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
strict = true

[tool.ruff]
line-length = 79
select = [
    "A",
    "ARG",
    "B",
    "C",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "T10",
    "UP",
    "W",
]
ignore = ["A003"]
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "S"]
